requirements.txt
numpy<2.0.0
soundcard
sounddevice
SpeechRecognition
google-generativeai
scipy
pyaudio